+++ index.php
@@ -18,11 +18,11 @@
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
         <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
         <link rel="icon" href="favicon.ico" type="image/gif" sizes="16x16">
-        <link rel="stylesheet" type="text/css" href="css/main.css">
-        <link rel="stylesheet" type="text/css" href="css/index/editing.css">
-        <link rel="stylesheet" type="text/css" href="css/navbar.css">
-        <link rel="stylesheet" type="text/css" href="css/index/container.css">
-        <link rel="stylesheet" type="text/css" href="css/colors.php">
+        <link rel="stylesheet" type="text/css" href="css/main.css?v=1.0">
+        <link rel="stylesheet" type="text/css" href="css/index/editing.css?v=1.0">
+        <link rel="stylesheet" type="text/css" href="css/navbar.css?v=1.0">
+        <link rel="stylesheet" type="text/css" href="css/index/container.css?v=1.0">
+        <link rel="stylesheet" type="text/css" href="css/colors.php?v=1.0">
         <link href="https://fonts.googleapis.com/icon?family=Material+Icons+Outlined" rel="stylesheet">
         <script src="https://code.jquery.com/jquery-3.5.1.min.js" integrity="sha256-9/aliU8dGd2tb6OSsuzixeV4y/faTqgFtohetphbbj0=" crossorigin="anonymous"></script>
         <title>Startpage</title>
@@ -49,7 +49,7 @@
                             $(".linkA").each(function() {
                                 var linkName = $(this).html();
                                 var linkHref = $(this).attr("href");
-                                var searchLink = '<a class="searchLink c3_cl" href="'+linkHref+'">'+linkName+'</a>';
+                                var searchLink = '<a class="searchLink c5_cl" href="'+linkHref+'">'+linkName+'</a>';
 
                                 if(linkName.toLowerCase().includes(input) || linkHref.toLowerCase().includes(input)) {
                                     $(searchLink).appendTo(".searchResults");
@@ -135,8 +135,8 @@
     </head>
     <body class="c1_bg">
         <nav class="topnav c3_bg">
-            <input type="search" class="searchBar c4_bg c3_cl" placeholder="Search">
-            <ul class="searchResults c4_bg c3_cl" style="display: none;">
+            <input type="search" class="searchBar c2_bg c3_cl" placeholder="Search">
+            <ul class="searchResults c4_bg" style="display: none;">
             </ul>
             <div class="navIconContainer">
                 <input class="editStatus" value="false" disabled hidden>