+++ php/loadColors.php
@@ -5,9 +5,10 @@
     $colorsSql = "SELECT * FROM colors WHERE user_link=$uid";
 
     $C1 = "#cdcdcd"; //BACKGROUND COLOR
-    $C2 = "#ffffff"; //TEXT COLOR
-    $C3 = "#071013"; //ACCENT COLOR 1
-    $C4 = "#ffffff"; //ACCENT COLOR 2
+    $C2 = "#ffffff"; //HEADER TEXT COLOR
+    $C3 = "#071013"; //HEADER BACKGROUND COLOR
+    $C4 = "#ffffff"; //CONTAINER BACKGROUND COLOR
+    $C5 = "#071013"; //CONTAINER TEXT COLOR
 
     if($result = mysqli_query($con, $colorsSql)) {
         while($row = mysqli_fetch_assoc($result)) {
@@ -15,10 +16,11 @@
             $user_id = $row["color_id"];
             $user_link = $row["user_link"];
 
-            $C1 = $row["color_1"]; //BACKGROUND COLOR
-            $C2 = $row["color_2"]; //TEXT COLOR
-            $C3 = $row["color_3"]; //ACCENT COLOR 1
-            $C4 = $row["color_4"]; //ACCENT COLOR 2
+            if(!empty($row["color_1"])){$C1 = $row["color_1"];}; //BACKGROUND COLOR
+            if(!empty($row["color_2"])){$C2 = $row["color_2"];}; //HEADER TEXT COLOR
+            if(!empty($row["color_3"])){$C3 = $row["color_3"];}; //HEADER BACKGROUND COLOR
+            if(!empty($row["color_4"])){$C4 = $row["color_4"];}; //CONTAINER BACKGROUND COLOR
+            if(!empty($row["color_5"])){$C5 = $row["color_5"];}; //CONTAINER TEXT COLOR
         }
     }
 ?>